:card_file_box: 실제 데이터베이스와 필요한 기능에 맞게 코드 수정
* Mapper: 데이터 베이스 실제 데이터에 맞게 변경
* Repository: 실제 기능 구현에 필요한 Query로 interface 구성 변경

Related to: #39

+++ application/src/main/java/com/triget/application/domain/accommodation/Accommodation.java
@@ -23,7 +23,7 @@ public class Accommodation {
     private String city;
     private String district;
     private String address;
-    private int price;
+    private float price;
     @Field("currency_code")
     private String currencyCode;
     private float rating;
@@ -34,7 +34,7 @@ public class Accommodation {
     @Builder
     public Accommodation(ObjectId _id, String bookingsId, String name, String thumbnail, String subcategory,
                          float longitude, float latitude, String city, String district, String address,
-                         int price, String currencyCode, float rating, int popularity, String detailUrl)
+                         float price, String currencyCode, float rating, int popularity, String detailUrl)
     {
         this._id = _id;
         this.bookingsId = bookingsId;
